+++ src/components/Media/Media.js
@@ -6,11 +6,30 @@ import CardMedia from "@material-ui/core/CardMedia";
 import Skeleton from "@material-ui/lab/Skeleton";
 import Fade from "@material-ui/core/Fade";
 import ThumbUpIcon from "@material-ui/icons/ThumbUp";
-import ThumbDownIcon from '@material-ui/icons/ThumbDown';
-import {Link} from 'prismic-reactjs';
-import {
-  PRISMIC_MEDIALIB_DOC
-} from "../../utils/constants";
+import ThumbDownIcon from "@material-ui/icons/ThumbDown";
+import CloseIcon from "@material-ui/icons/Close";
+import { Link } from "prismic-reactjs";
+import { PRISMIC_MEDIALIB_DOC } from "../../utils/constants";
+
+const MediaTitleComp = ({children, setModal}) => {
+  return (
+    <div
+      style={{
+        display: "flex",
+        justifyContent: "space-between",
+        alignItems: "center",
+        minHeight: "60px"
+      }}
+    >
+      {children}
+      <CloseIcon
+        color="primary"
+        onClick={() => {return setModal({ show: false, name: "" })}}
+      />
+    </div>
+  );
+};
+
 
 const Media = props => {
   // All you need for loggin
@@ -20,10 +39,16 @@ const Media = props => {
   const iframeRef = useRef(null);
   const { cmsContents, setModal } = useContext(UserContext);
   useEffect(() => {
-    if (cmsContents  && cmsContents.data  && cmsContents.data.hasOwnProperty(PRISMIC_MEDIALIB_DOC)) {
-      console.log(cmsContents.data[PRISMIC_MEDIALIB_DOC].edges[0].node.mediaTitle);
+    if (
+      cmsContents &&
+      cmsContents.data &&
+      cmsContents.data.hasOwnProperty(PRISMIC_MEDIALIB_DOC)
+    ) {
+      console.log(
+        cmsContents.data[PRISMIC_MEDIALIB_DOC].edges[0].node.mediaTitle
+      );
     }
-  }, [cmsContents])
+  }, [cmsContents]);
   useEffect(() => {
     setLogs({
       action: "MEDIA",
@@ -36,14 +61,10 @@ const Media = props => {
     visibility: "hidden",
     position: "absolute"
   });
-
+  
   const initialState = {
     description: "",
-    title: (
-      <div style={{ minHeight: "48px" }}>
-        <Skeleton width="100%" />
-      </div>
-    ),
+    title: <MediaTitleComp setModal={setModal}><Skeleton width="100%" /></MediaTitleComp>,
     primaryButton: <ThumbUpIcon />,
     secondaryButton: <ThumbDownIcon />,
     error: false,
@@ -55,11 +76,11 @@ const Media = props => {
       setIframestyles({ visibility: "visible", position: "static" });
       setModalConfig({
         ...modalConfig,
-        title: "Staying Positive After Hardships"
+        title: <MediaTitleComp setModal={setModal}><span>Staying Positive After Hardships</span></MediaTitleComp>
       });
     }
   }, [iframeloading, modalConfig]);
-  
+
   // useEffect(() => {
   //   if (methods.hasOwnProperty("error")) {
   //     setModalConfig({
@@ -76,21 +97,20 @@ const Media = props => {
   const handlePrimary = () => {
     setModalConfig({ ...modalConfig, loading: true });
 
-      setLogs(() => {
-        setModal({ show: false, name: "" });
-        return {
-          action: "Likes",
-          message: `just liked the media content`
-        };
-      });
-      
+    setLogs(() => {
+      setModal({ show: false, name: "" });
+      return {
+        action: "Likes",
+        message: `just liked the media content`
+      };
+    });
   };
 
   const handleSecondary = () => {
     setLogs({
-        action: "Sucks",
-        message: `just Disliked the media content`
-      });
+      action: "Sucks",
+      message: `just Disliked the media content`
+    });
   };
   return (
     <DialogModal
@@ -106,8 +126,14 @@ const Media = props => {
       handleSecondaryAction={() => handleSecondary()}
       loading={modalConfig.loading}
       fullWidth={true}
+      fullScreen={true}
+      contentContainerStyle={{
+        justifyContent: "center",
+        alignItems: "center",
+        display: "flex"
+      }}
     >
-      <div style={{ minHeight: "155px" }}>
+      <div style={{ minHeight: "155px", width: "100%" }}>
         {iframeloading ? (
           <div>
             <Skeleton variant="rect" width={"100%"} height={145} />