+++ CHANGELOG.md
@@ -38,7 +38,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [3.2.0] - 2019-11-..
 ### Added
-- Media Playback at homescreen (Prayertimes screen)
+- Media Playback at homescreen (Prayertimes screen) with fullscreen paper
 - Hook up with Prismic
 
 ## [3.1.0] - 2019-10-28